*: nixfmt

+++ users/bobby285271/default.nix
@@ -21,19 +21,27 @@
     ];
   };
 
-  security.sudo.extraRules = [{
-    users = [ "bobby285271" ];
-    commands = [{
-      command = "ALL";
-      options = [ "NOPASSWD" ];
-    }];
-  }];
+  security.sudo.extraRules = [
+    {
+      users = [ "bobby285271" ];
+      commands = [
+        {
+          command = "ALL";
+          options = [ "NOPASSWD" ];
+        }
+      ];
+    }
+  ];
 
-  system.activationScripts.setUserImages.text = if config.services.xserver.desktopManager.cinnamon.enable then ''
-    rm -f /var/lib/AccountsService/icons/bobby285271
-  '' else ''
-    if test -f /var/lib/AccountsService/icons/bobby285271.default; then
-      cp -f /var/lib/AccountsService/icons/bobby285271{.default,}
-    fi
-  '';
+  system.activationScripts.setUserImages.text =
+    if config.services.xserver.desktopManager.cinnamon.enable then
+      ''
+        rm -f /var/lib/AccountsService/icons/bobby285271
+      ''
+    else
+      ''
+        if test -f /var/lib/AccountsService/icons/bobby285271.default; then
+          cp -f /var/lib/AccountsService/icons/bobby285271{.default,}
+        fi
+      '';
 }
